fix: save previous changes
