+++ src/services/decks/decks.ts
@@ -1,4 +1,5 @@
 import { baseApi } from '@/services/base-api.ts'
+import { CreateDeckArgs, Deck, DecksResponse, GetDecksArgs } from '@/services/decks/types'
 
 const decksApi = baseApi.injectEndpoints({
   endpoints: builder => {
@@ -9,7 +10,7 @@ const decksApi = baseApi.injectEndpoints({
         },
         providesTags: ['Decks'],
       }),
-      createDeck: builder.mutation<Deck, { name: string }>({
+      createDeck: builder.mutation<Deck, CreateDeckArgs>({
         query: ({ name }) => {
           return { url: `v1/decks`, method: 'POST', body: { name } }
         },
@@ -20,47 +21,3 @@ const decksApi = baseApi.injectEndpoints({
 })
 
 export const { useGetDecksQuery, useCreateDeckMutation } = decksApi
-
-type GetDecksArgs = {
-  minCardsCount?: number
-  maxCardsCount?: number
-  name?: string
-  authorId?: string
-  orderBy?: string
-  currentPage?: number
-  itemsPerPage?: number
-}
-
-export interface Pagination {
-  totalPages: number
-  currentPage: number
-  itemsPerPage: number
-  totalItems: number
-}
-
-export interface Author {
-  id: string
-  name: string
-}
-
-export interface Deck {
-  id: string
-  userId: string
-  name: string
-  isPrivate: boolean
-  shots: number
-  cover?: string | null
-  rating: number
-  isDeleted?: boolean | null
-  isBlocked?: boolean | null
-  created: string
-  updated: string
-  cardsCount: number
-  author: Author
-}
-
-export interface DecksResponse {
-  maxCardsCount: number
-  pagination: Pagination
-  items: Deck[]
-}
