+++ src/services/types.ts
@@ -1,6 +1,13 @@
-import { Pagination } from '@/services/decks/types.ts'
-
 export type PaginatedEntity<T> = {
   pagination: Pagination
   items: T[]
 }
+
+export interface Pagination {
+  totalPages: number
+  currentPage: number
+  itemsPerPage: number
+  totalItems: number
+}
+
+export type PaginatedRequest = Partial<Pick<Pagination, 'itemsPerPage' | 'currentPage'>>
